requirements.txt
pyro-ppl==1.3.1
torch==1.4.0+cpu
scipy==1.2.1
tqdm==4.45.0